WIP: Refactoring, async reporting of results. Compiles but not at all tested.
