+++ src/analysis/prusti_line_consumer.ts
@@ -1,0 +1,5 @@
+import * as vscode from "vscode";
+
+export interface PrustiLineConsumer extends vscode.Disposable {
+    tryConsumeLine: (line: string, isCrate: boolean, programPath: string) => boolean
+}
