+++ src/analysis/verificationResult.ts
@@ -2,10 +2,6 @@ import * as util from "./../util";
 import * as vscode from "vscode";
 import * as path from "path";
 
-interface VerificationInfoRaw {
-    result_list: VerificationResultRaw[]
-}
-
 interface VerificationResultRaw {
     item_name: string,
     success: boolean,
@@ -20,7 +16,6 @@ export interface VerificationResult {
     cached: boolean,
 }
 
-
 // currently the value given for itemName is "filename_methodpath"
 // should this be done in rust?
 function splitName(name: string) : [string, string] {
@@ -32,46 +27,29 @@ function splitName(name: string) : [string, string] {
     return [filename, methodPath]
 }
 
-function transformVerificationResult(rawResults: VerificationResultRaw[], isCrate: boolean, rootPath: string) : VerificationResult[] {
-    let dirPath: string;
-    if (isCrate) {
-        dirPath = rootPath;
-    } else {
-        dirPath = path.dirname(rootPath) + "/";
-    }
+function transformVerificationResult(rawRes: VerificationResultRaw, isCrate: boolean, rootPath: string) : VerificationResult {
+    let dirPath = isCrate ? rootPath : path.dirname(rootPath) + "/";
 
-    let results: VerificationResult[] = [];
-    rawResults.forEach((rawRes) => {
-        let [_fileName, methodPath] = splitName(rawRes.item_name);
-        // we realized this fileName is not useful, for crates it's always main.rs
-        let res = {
-            methodName: methodPath,
-            success: rawRes.success,
-            time_ms: rawRes.time_ms,
-            cached: rawRes.cached,
-        };
-        results.push(res);
-    });
-    return results;
+    let [_fileName, methodPath] = splitName(rawRes.item_name);
+    // we realized this fileName is not useful, for crates it's always main.rs
+    let res = {
+        methodName: methodPath,
+        success: rawRes.success,
+        time_ms: rawRes.time_ms,
+        cached: rawRes.cached,
+    };
+    return res;
 }
 
-export function parseVerificationResult(output: string, isCrate: boolean, rootPath: string): VerificationResult[] {
-    let token = "VerificationInfo ";
-    let len = token.length;
-    for (const line of output.split("\n")) {
-        if (!line.startsWith(token)) {
-            continue;
-        }
-
-        let rawResult = JSON.parse(line.substring(len)) as VerificationInfoRaw;
-        if (rawResult.result_list !== undefined) {
-            util.log("Parsed verification summary. Found"
-                    + rawResult.result_list.length
-                    + " verification-results.");
-            return transformVerificationResult(rawResult.result_list, isCrate, rootPath);
-
-        }
+export function parseVerificationResult(line: string, isCrate: boolean, rootPath: string): VerificationResult | undefined {
+    const msg = util.getMessage(line, isCrate);
+    if (msg === undefined) {
+        return false;
     }
-    return []
+    const token = "ide_verification_result";
+    if (!msg.message.startsWith(token)) {
+        return undefined;
+    }
+    let rawResult = JSON.parse(line.substring(token.length)) as VerificationResultRaw;
+    return transformVerificationResult(rawResult, isCrate, rootPath);
 }
-
