+++ src/util.ts
@@ -1,8 +1,8 @@
 import * as childProcess from "child_process";
 import * as vscode from "vscode";
-import * as fs from "fs";
-import * as path from "path";
 import * as treeKill from "tree-kill";
+import { CargoMessage, Message } from "./analysis/diagnostics"
+import { projects } from "./projects";
 
 export function userInfo(message: string, statusBar?: vscode.StatusBarItem): void {
     log(message);
@@ -191,54 +191,72 @@ export function spawn(
     });
 }
 
-export class Project {
-    readonly path;
-
-    public constructor(_path: string) {
-        this.path = _path;
-    }
-
-    public hasRootFile(fileName: string): Promise<boolean> {
-        const filePath = path.join(this.path, fileName);
-        return new Promise(resolve => {
-            fs.access(filePath, fs.constants.F_OK, (err) => resolve(err === null));
-        });
-    }
-}
-
-export class ProjectList {
-    public constructor(readonly projects: Project[]) {}
-
-    public isEmpty(): boolean {
-        return this.projects.length === 0;
-    }
-
-    public getParent(file: string): Project | undefined {
-        let result: Project | undefined;
-        // Find the last (innermost) project that contains the file.
-        for (const project of this.projects) {
-            if (file.startsWith(project.path)) {
-                result = project;
-            }
-        }
-        return result;
-    }
-}
-
 /**
- * Find all projects in the workspace that contain a Cargo.toml file.
- *
- * @returns A project list.
+ * Given a range, possibly spanning multiple lines this function will return a range
+ * that includes all of the last line. The purpose of this is that decorators
+ * that are displayed "behind" this range, will not be in the middle of some text
  */
-export async function findProjects(): Promise<ProjectList> {
-    const projects: Project[] = [];
-    (await vscode.workspace.findFiles("**/Cargo.toml")).forEach((uri: vscode.Uri) => {
-        projects.push(new Project(uri.fsPath.replace(/[/\\]?Cargo\.toml$/, "")));
-    });
-    projects.sort((a, b) => {
-        if (a.path > b.path) { return 1; }
-        if (a.path < b.path) { return -1; }
-        return 0;
-    });
-    return new ProjectList(projects);
+export function full_line_range(range: vscode.Range): vscode.Range {
+    let position = new vscode.Position(range.start.line, range.start.character);
+    let position_test = new vscode.Position(range.start.line, Number.MAX_SAFE_INTEGER);
+
+    return new vscode.Range(position, position_test)
+}
+
+// we could implement some more thorough checking, but for now, this suffices
+export function isCargoMessage(msg: Message|CargoMessage): msg is CargoMessage {
+    return ((msg as CargoMessage).target !== undefined);
+}
+
+export function isValidCargoMessage(msg: CargoMessage): boolean {
+    return (msg.message !== undefined && isValidRustcMessage(msg.message));
+}
+
+export function isValidRustcMessage(msg: Message) {
+    // TODO: remove FakeError once fixed
+    return (msg.message !== undefined && msg.message !== "[Prusti: FakeError]");
+}
+
+export function getCargoMessage(line: string): CargoMessage|undefined {
+    const msg = JSON.parse(line) as CargoMessage;
+    if (!isValidCargoMessage(msg)) {
+        return undefined;
+    }
+    return msg;
+}
+
+export function getRustcMessage(line: string): Message|undefined {
+    const msg = JSON.parse(line) as Message;
+    if (!isValidRustcMessage(msg)) {
+        return undefined;
+    }
+    return msg;
+}
+
+export function getMessage(line: string, isCrate: boolean): Message|undefined {
+    if (isCrate) {
+        const cargoMessage = getCargoMessage(line);
+        if (cargoMessage !== undefined) {
+            return cargoMessage.message;
+        }
+        return undefined;
+    } else {
+        const rustcMessage = getRustcMessage(line);
+        return rustcMessage;
+    }
+}
+
+/** Either returns the path to the root of the crate containing
+* the file (at filePath), or just filePath itself, if it's
+* a standalone file
+*/
+export function getRootPath(filePath: string): string {
+    let res;
+    let parent = projects.getParent(filePath);
+    if (parent !== undefined) {
+        res = parent.path + "/";
+    } else {
+        res = filePath;
+    }
+    return res;
 }
